# docs/swagger.yaml
definitions:
  model.Application:
    properties:
      active_manifest_id:
        type: string
      active_manifest_name:
        type: string
      config_maps:
        items:
          $ref: '#/definitions/model.ConfigMap'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      envs:
        additionalProperties:
          items:
            $ref: '#/definitions/model.EnvVar'
          type: array
        type: object
      id:
        type: string
      internet:
        $ref: '#/definitions/model.Internet'
      name:
        type: string
      project_name:
        type: string
      replica:
        type: integer
      repo_url:
        type: string
      service:
        $ref: '#/definitions/model.Service'
      status:
        description: 当前状态（来自 Job 的结果）
        type: string
      type:
        $ref: '#/definitions/model.ReleaseType'
      updated_at:
        type: string
    type: object
  model.ConfigMap:
    properties:
      files_path:
        additionalProperties:
          type: string
        type: object
      mount_path:
        type: string
      name:
        type: string
    type: object
  model.Configuration:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.EnvVar:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  model.File:
    properties:
      content:
        type: string
      name:
        type: string
    type: object
  model.Internet:
    enum:
    - internal
    - external
    type: string
    x-enum-varnames:
    - Internal
    - External
  model.Job:
    properties:
      application_id:
        type: string
      application_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      env:
        type: string
      id:
        type: string
      manifest_id:
        type: string
      manifest_name:
        type: string
      project_name:
        type: string
      status:
        $ref: '#/definitions/model.JobStatus'
      type:
        type: string
      updated_at:
        type: string
    type: object
  model.JobStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    - RollingBack
    - RolledBack
    - Syncing
    - SyncFailed
    type: string
    x-enum-varnames:
    - JobPending
    - JobRunning
    - JobSucceeded
    - JobFailed
    - JobRollingBack
    - JobRolledBack
    - JobSyncing
    - JobSyncFailed
  model.Manifest:
    properties:
      application_id:
        description: 关联 Application
        type: string
      application_name:
        type: string
      branch:
        description: git branch
        type: string
      commit_hash:
        type: string
      config_maps:
        items:
          $ref: '#/definitions/model.ConfigMap'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      digest:
        type: string
      envs:
        additionalProperties:
          items:
            $ref: '#/definitions/model.EnvVar'
          type: array
        type: object
      git_repo:
        description: 对应 Application repo
        type: string
      id:
        type: string
      internet:
        $ref: '#/definitions/model.Internet'
      name:
        type: string
      pipeline_id:
        description: Tekton PipelineRun ID
        type: string
      replica:
        type: integer
      service:
        $ref: '#/definitions/model.Service'
      status:
        allOf:
        - $ref: '#/definitions/model.ManifestStatus'
        description: running, success, failed
      steps:
        description: 每个步骤状态
        items:
          $ref: '#/definitions/model.ManifestStep'
        type: array
      type:
        $ref: '#/definitions/model.ReleaseType'
      updated_at:
        type: string
    type: object
  model.ManifestStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    type: string
    x-enum-varnames:
    - ManifestPending
    - ManifestRunning
    - ManifestSucceeded
    - ManifestFailed
  model.ManifestStep:
    properties:
      end_time:
        type: string
      message:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/model.StepStatus'
      task_name:
        type: string
      task_run:
        type: string
    type: object
  model.PatchManifestRequest:
    properties:
      commit_hash:
        description: Git commit hash used to generate this manifest
        type: string
      digest:
        description: Image digest generated by CI pipeline (e.g. sha256:xxx)
        type: string
    type: object
  model.Port:
    properties:
      name:
        type: string
      port:
        type: integer
      target_port:
        type: integer
    type: object
  model.ReleaseType:
    enum:
    - normal
    - canary
    - blue-green
    type: string
    x-enum-varnames:
    - Normal
    - Canary
    - BlueGreen
  model.Service:
    properties:
      ports:
        items:
          $ref: '#/definitions/model.Port'
        type: array
    type: object
  model.StepStatus:
    enum:
    - Pending
    - Running
    - Succeeded
    - Failed
    type: string
    x-enum-varnames:
    - StepPending
    - StepRunning
    - StepSucceeded
    - StepFailed
  pkg_api.UpdateActiveManifestRequest:
    properties:
      manifest_id:
        type: string
    required:
    - manifest_id
    type: object
info:
  contact:
    email: devflow@example.com
    name: DevFlow Team
    url: http://devflow.example.com
  description: DevFlow CD 平台 REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://devflow.example.com/terms/
  title: DevFlow CD Platform API
  version: "1.0"
paths:
  /api/v1/applications:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Application'
            type: array
      summary: 获取应用列表
      tags:
      - Application
    post:
      consumes:
      - application/json
      description: 创建一个新的应用
      parameters:
      - description: Application Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Application'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建应用
      tags:
      - Application
  /api/v1/applications/{id}:
    delete:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除应用
      tags:
      - Application
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Application'
      summary: 获取应用
      tags:
      - Application
    put:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Application Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Application'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用
      tags:
      - Application
  /api/v1/applications/{id}/active_manifest:
    patch:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Active Manifest Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_api.UpdateActiveManifestRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新应用的 Active Manifest
      tags:
      - Application
  /api/v1/configurations:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Configuration'
            type: array
      summary: 获取配置列表
      tags:
      - Configuration
    post:
      consumes:
      - application/json
      description: 创建一个新的配置
      parameters:
      - description: Configuration Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Configuration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建配置
      tags:
      - Configuration
  /api/v1/configurations/{id}:
    delete:
      parameters:
      - description: Configuration ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除配置
      tags:
      - Configuration
    get:
      parameters:
      - description: Configuration ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Configuration'
      summary: 获取配置
      tags:
      - Configuration
    put:
      parameters:
      - description: Configuration ID
        in: path
        name: id
        required: true
        type: string
      - description: Configuration Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Configuration'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新配置
      tags:
      - Configuration
  /api/v1/jobs:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Job'
            type: array
      summary: 获取Job列表
      tags:
      - Job
    post:
      consumes:
      - application/json
      description: 创建一个新的Job
      parameters:
      - description: Job Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Job'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建Job
      tags:
      - Job
  /api/v1/jobs/{id}:
    delete:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除Job
      tags:
      - Job
    get:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Job'
      summary: 获取Job
      tags:
      - Job
    put:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Job Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Job'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新Job
      tags:
      - Job
  /api/v1/manifests:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Manifest'
            type: array
      summary: 获取应用列表
      tags:
      - Manifest
    post:
      consumes:
      - application/json
      description: 根据 Manifest 创建 Manifest，自动生成名称
      parameters:
      - description: Manifest 数据（branch 必填）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Manifest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Manifest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建 Manifest
      tags:
      - Manifest
  /api/v1/manifests/{id}:
    get:
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Manifest'
      summary: 获取应用
      tags:
      - Manifest
    patch:
      consumes:
      - application/json
      description: 部分更新 Manifest（仅支持 digest / commit_hash）
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: string
      - description: Patch 数据
        in: body
        name: data
        schema:
          $ref: '#/definitions/model.PatchManifestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch Manifest
      tags:
      - Manifest
schemes:
- http
- https
swagger: "2.0"